+++ flutter_mediapipe_vision_platform_interface/lib/models.dart
@@ -1,18 +1,27 @@
 import 'dart:ui';
 
+import 'package:json_annotation/json_annotation.dart';
+
+part 'models.g.dart';
+
+@JsonSerializable()
 class PoseLandmarkerResult {
   final List<List<NormalizedLandmark>> landmarks;
 
   const PoseLandmarkerResult.empty() : landmarks = const [];
 
-  const PoseLandmarkerResult({
-    required this.landmarks,
-  });
+  const PoseLandmarkerResult({required this.landmarks});
 
   @override
   String toString() => landmarks.toString();
+
+  Map<String, dynamic> toJson() => _$PoseLandmarkerResultToJson(this);
+
+  factory PoseLandmarkerResult.fromJson(Map<String, dynamic> map) =>
+      _$PoseLandmarkerResultFromJson(map);
 }
 
+@JsonSerializable()
 class NormalizedLandmark {
   final double x;
   final double y;
@@ -30,4 +39,9 @@ class NormalizedLandmark {
 
   @override
   String toString() => '[$x, $y, $z]';
+
+  Map<String, dynamic> toJson() => _$NormalizedLandmarkToJson(this);
+
+  factory NormalizedLandmark.fromJson(Map<String, dynamic> map) =>
+      _$NormalizedLandmarkFromJson(map);
 }