+++ flutter_mediapipe_vision_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.1
+
+* Make `PoseLandmarkerResult` and `NormalizedLandmark` serializable to JSON.
+
 ## 0.2.0
 
 * Added `z` and `visibility` to `NormalizedLandmark`.